website: move resources to community
