+++ website/pages/docs/who-uses-nomad.mdx
@@ -1,5 +1,5 @@
 ---
-layout: intro
+layout: docs
 page_title: Who Uses Nomad
 sidebar_title: Who Uses Nomad
 description: >-
